% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.overarching_SuperLearner.R
\name{predict.overarching_SuperLearner}
\alias{predict.overarching_SuperLearner}
\title{Predict method for SequentialSuperLearner object}
\usage{
\method{predict}{overarching_SuperLearner}(object, newdata, ...)
}
\arguments{
\item{object}{Fitted object from \code{overarching_SuperLearner}}

\item{newdata}{New X values for prediction}

\item{...}{Not used.}
}
\value{
\item{base_learners_predictions}{Predicted values for each base algorithm in library}
\item{meta_learners_predictions}{Predicted values for each meta-algorithm in library}
\item{overarching_predictions}{ Predicted values from the overarching super learner fit}
}
\description{
Obtains predictions on a new data set from a SequentialSuperLearner fit.
}
\details{
Each algorithm in the Super Learner library needs to have a
corresponding prediction function with the ``predict.'' prefixed onto the
algorithm name (e.g. \code{predict.SL.glm} for \code{SL.glm}).
}
\references{
van  der Laan, M. J.,  Polley, E. C. and Hubbard,  A. E. (2008)
Super Learner,  \emph{Statistical Applications of Genetics  and Molecular
Biology}, \bold{6}, article 25.

Ecoto, G.,  Bibaut, A. and  Chambaz, A. (2021) One-step  ahead sequential
Super Learning from  short times series of many  slightly dependent data,
and     anticipating      the     cost     of      natural     disasters,
\url{https://arxiv.org/abs/2107.13291}.
}
\seealso{
\code{\link{SuperLearner}}
}
\author{
Geoffrey   Ecoto  \email{geoffrey.ecoto@gmail.com}  and  Antoine
Chambaz \email{antoine.chambaz@u-paris.fr}
}
\keyword{models}
