% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overarching_SuperLearner.R
\docType{package}
\name{SequentialSuperLearner}
\alias{SequentialSuperLearner}
\title{SequentialSuperLearner: Implements the overarching sequential super learning algorithm}
\description{
The overarching sequential super learning algorithm is a variant of the super
learning algorithm. Designed to learn from times series, it sequentially identifies
the best algorithm in a library, or the best combination of algorithms in the library,
where the said library consists of several super learners.
}
